Study program ID update

+++ src/main/java/lt/sdacademy/university/models/dto/StudyProgram.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class StudyProgram {
 
-    private Integer id;
+    private Long id;
 
     private String title;
 
@@ -13,17 +13,17 @@ public class StudyProgram {
     public StudyProgram() {
     }
 
-    public StudyProgram(Integer id, String title, List<String> modules) {
+    public StudyProgram(Long id, String title, List<String> modules) {
         this.id = id;
         this.title = title;
         this.modules = modules;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
